INCS-5657 created startConnectionData variable to pass into StartConnection component
